requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
gunicorn==21.2.0
python-dotenv==1.0.1
requests==2.32.3
APScheduler==3.10.4
Werkzeug==2.3.8
PyMySQL==1.1.0
cryptography==41.0.0